requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
python-dotenv==0.19.0
pandas==1.3.3
openpyxl==3.0.9
requests==2.26.0
Werkzeug==2.0.1 